[#22]: added chosen drafted expansion; added constants:

+++ src/main/java/it/capozxvii/terraformingmars/model/dto/DraftingDto.java
@@ -22,5 +22,5 @@ import lombok.experimental.SuperBuilder;
 public class DraftingDto extends AbstractDto implements Serializable {
     private PlayerDto player;
     private Long championshipId;
-    private Map<Integer, CorporationsExpansionPairDto> draftings;
+    private Map<Long, CorporationsExpansionPairDto> draftings;
 }
